refactor: move client port resolution to a light leaf module (#2933)
mcp_core resolved its gateway API port through a function-local import of
cli_server, whose top-level import graph (frontend build, service
controllers, preflight, embeddings) is far heavier than the MCP stdio
server's own. Move the resolution chain (resolve_client_port,
resolve_client_port_ex, _config_url_port, _marker_port,
_gateway_owns_port, _is_kirocrew_process, plus the argv-classifier
helpers they depend on) into kiro_crew.port_resolution, a leaf module
both consumers import at module scope. Pure move, no logic edits.

cli_server re-exports every moved name, so its public surface is
unchanged and the existing tests that patch these symbols in the
cli_server namespace keep working: chain-internal calls resolve their
callee through the patched namespace via a sys.modules lookup
(_patchable) whenever cli_server is loaded, falling back to the leaf's
own globals in processes that never load it. A regression test asserts
in a fresh interpreter that importing port_resolution or mcp_core does
not pull in cli_server.

Closes #2933
