+++ src/kiro_crew/cli_server.py
@@ -7,14 +7,12 @@ import http.client
 import json
 import logging
 import os
-import shlex
 import shutil
 import signal
 import subprocess
 import sys
 import time
 import urllib.error
-import urllib.parse
 import urllib.request
 from pathlib import Path
 
@@ -22,7 +20,6 @@ from kiro_crew import __version__, platform_compat
 from kiro_crew.beacon import is_default_home
 from kiro_crew.config import KiroCrewConfig
 from kiro_crew.config.loader import (
-    _DEFAULT_PORT,
     _session_work_dir,
     build_provider_factory,
     config_dir,
@@ -34,7 +31,6 @@ from kiro_crew.dashboard import tailnet_serve
 from kiro_crew.dashboard.handlers.core import DASHBOARD_HTML_NOT_FOUND_MARKER
 from kiro_crew.dashboard.origin import (
     dashboard_origin,
-    parse_dashboard_url,
     resolve_dashboard_host,
 )
 from kiro_crew.dashboard.tailnet import is_governance_pinned_off, tailnet_origin
@@ -52,6 +48,26 @@ from kiro_crew.instances import run_marker
 from kiro_crew.learn import LessonStore
 from kiro_crew.loopback_http import loopback_urlopen
 from kiro_crew.memory import MemoryStore
+
+# Client-side port resolution lives in kiro_crew.port_resolution, a light leaf
+# module the MCP stdio server can import without paying for this module's
+# import graph. Re-exported here (not merely used) because this namespace is
+# the historical public surface: external callers and a large body of tests
+# import and patch these names as ``kiro_crew.cli_server.<name>``, and the
+# chain's internal calls deliberately resolve through this namespace whenever
+# it is loaded (see port_resolution._patchable) so those patches keep
+# intercepting.
+from kiro_crew.port_resolution import (  # noqa: F401
+    _KIROCREW_SERVER_SUBCOMMANDS,
+    _args_look_like_kirocrew,
+    _basename_stem,
+    _config_url_port,
+    _gateway_owns_port,
+    _is_kirocrew_process,
+    _marker_port,
+    resolve_client_port,
+    resolve_client_port_ex,
+)
 from kiro_crew.preflight import run_preflight_checks
 from kiro_crew.sel import sel
 from kiro_crew.service import controller as service_controller
@@ -79,224 +95,6 @@ from kiro_crew.vector_memory import VectorMemoryStore
 # SPA's per-origin localStorage is keyed on that host, so emitting a different
 # origin would make every dashboard setting appear reset.
 _CLI_LOOPBACK = "127.0.0.1"
-
-
-def _config_url_port() -> int | None:
-    """Port explicitly named by ``dashboard.url``, or ``None``.
-
-    Distinct from :func:`parse_dashboard_url`, which substitutes
-    ``_DEFAULT_PORT`` for a portless URL (``http://my.host``). That substitution
-    is right for the *server* (it must bind something) but wrong for a client:
-    it would report "config says 5476" and short-circuit the run-marker
-    fallback below, even though the user never named a port. So detect the
-    explicit case and let a portless URL fall through.
-    """
-    try:
-        cfg = KiroCrewConfig.load()
-        url = cfg.dashboard.url or ""
-    except Exception:
-        # Config load failures must not break client commands.
-        return None
-    if not isinstance(url, str):
-        # ``dashboard.url`` is user-editable JSON and core installs may lack
-        # jsonschema, so the value can be any type (``"url": 123``). urlparse
-        # raises TypeError on a non-str, which is NOT a ValueError — without
-        # this guard a bad config type would crash every client command.
-        logging.getLogger(__name__).warning(
-            "Ignoring non-string dashboard.url of type %s", type(url).__name__
-        )
-        return None
-    if not url:
-        return None
-    try:
-        _, port = parse_dashboard_url(url)
-        # parse_dashboard_url already normalises the scheme, tolerates malformed
-        # URLs and applies the KIROCREW_PORT override; re-split only to learn
-        # whether the port was written down or defaulted in.
-        explicit = urllib.parse.urlsplit(url if "://" in url else f"http://{url}").port
-    except (TypeError, ValueError):
-        return None
-    return port if explicit is not None else None
-
-
-def _gateway_owns_port(port: int) -> bool:
-    """True only when *this user's* gateway process is listening on *port*.
-
-    Reachability is not enough to trust a discovered port. Client commands hand
-    the local secret to whatever answers (``_token`` and ``_logout`` send
-    ``X-Local-Secret``), and ``clear_marker`` runs only on graceful shutdown —
-    so a crashed gateway leaves a marker naming a port some unrelated process
-    may since have bound. A bare TCP connect would walk the secret straight into
-    that process, which could then mint owner tokens against the real gateway.
-
-    A command-line check is not enough either: argv is attacker-chosen, so a
-    listener launched as ``/tmp/kirocrew gateway`` would pass it. The proof used
-    here is an identity the attacker cannot forge, in three parts:
-
-    1. **Recorded pid** — ``run_marker.read_pid(port)`` reads the sidecar the
-       gateway wrote at ``0600`` inside the ``0700`` ``run/`` dir. Another local
-       user cannot write it, so they cannot nominate a process of theirs.
-    2. **Holds the port** — that pid must be among
-       ``platform_compat.find_listening_pids(port)``. This is what makes a stale
-       recorded pid harmless: it has to actually hold the port we are about to
-       send the secret to.
-    3. **Owned by us, and ours** — the pid's uid must equal the caller's
-       (``process_owner_uid``), and its argv must look like a gateway. The uid
-       check is what closes pid *recycling* into a foreign user's process; argv
-       remains only as defense in depth, never as the sole proof.
-
-    A same-user attacker is out of scope by construction: they can already read
-    ``.local_secret`` (mode ``0600``, their own uid), so nothing here can be an
-    escalation for them. The boundary this closes is a *different* local user.
-
-    **Fails closed** at every step: no sidecar, no recorded pid, a pid that does
-    not hold the port, an unresolvable uid, a missing lookup tool
-    (``find_listening_pids`` folds that into an empty list) or a throwing one —
-    all deny, and discovery is skipped in favour of the documented default.
-    ``--port`` and ``KIROCREW_PORT`` remain available on such hosts.
-
-    **Non-POSIX denies outright.** ``process_owner_uid`` cannot report an owner
-    on Windows, and a home that is writable by another user (a shared or
-    misconfigured ``KIROCREW_HOME``) would let them replace both the marker and
-    the sidecar with a forged listener — the file-permission argument that
-    carries step 1 is exactly what stops holding there. Rather than trust
-    steps 1-2 alone, discovery is skipped: Windows users keep ``--port`` /
-    ``KIROCREW_PORT``, which is precisely where they were before this fallback
-    existed, so nothing regresses. This is the one place the feature is
-    deliberately unavailable rather than approximated.
-    """
-    if not platform_compat.IS_POSIX:
-        return False
-    recorded = run_marker.read_pid(port)
-    if recorded is None:
-        return False
-    try:
-        pids = platform_compat.find_listening_pids(port)
-    except Exception:
-        return False
-    if recorded not in pids:
-        return False
-    owner = platform_compat.process_owner_uid(recorded)
-    if owner is None or owner != os.getuid():
-        return False
-    return _is_kirocrew_process(recorded)
-
-
-def _marker_port() -> int | None:
-    """Port of the sole gateway-owned run-marker, or ``None``.
-
-    Zero-configuration discovery for the common single-gateway box: the gateway
-    already advertises itself by writing ``<data-home>/run/gateway-<port>.bin``
-    (see :mod:`kiro_crew.instances.run_marker`), so a client with no ``--port``,
-    no ``KIROCREW_PORT`` and no port in ``dashboard.url`` can read that instead
-    of assuming 5476 and connecting to a dead port.
-
-    Two guards keep this from being a guess:
-
-    * **Ownership.** Only ports where a verified KiroCrew gateway process is
-      listening count (:func:`_gateway_owns_port`); a stale marker, or one whose
-      port has been taken over by an unrelated process, is discarded.
-    * **Ambiguity.** With several gateways up there is no basis to pick one, so
-      this refuses (returns ``None``, landing on the documented default) and
-      tells the user on stderr which ports it saw and how to name one.
-    """
-    try:
-        candidates = run_marker.marker_ports()
-    except Exception:
-        return None
-    if not candidates:
-        return None
-    owned = [p for p in candidates if _gateway_owns_port(p)]
-    if len(owned) == 1:
-        return owned[0]
-    if len(owned) > 1:
-        print(
-            f"⚠️  Multiple gateways are running (ports {', '.join(str(p) for p in owned)}); "
-            f"not guessing which one you meant — using {_DEFAULT_PORT}. "
-            "Pass --port or set KIROCREW_PORT to target a specific gateway.",
-            file=sys.stderr,
-        )
-    return None
-
-
-def resolve_client_port(cli_port: int | None) -> int:
-    """Return the dashboard port a *client* CLI command (token/status/logout/stop)
-    should talk to.
-
-    Resolution order:
-
-    1. Explicit ``--port`` CLI flag if the user passed one (``cli_port`` is not ``None``).
-    2. ``KIROCREW_PORT`` env var if set to a valid integer. Deliberately ABOVE
-       the bound-port export: an explicitly-set ``KIROCREW_PORT`` is how a
-       caller retargets a child at a DIFFERENT gateway — ``pod exec`` builds a
-       client env with ``KIROCREW_PORT=<pod-port>`` precisely so the command
-       reaches the pod, while the inherited ``KIROCREW_BOUND_PORT`` still
-       names the spawning LIVE gateway. Ranking the bound value higher would
-       walk pod ``token``/``status``/``logout`` (and the local secret they
-       carry) into the live gateway — a cross-plane isolation break.
-    3. ``KIROCREW_BOUND_PORT`` env var if set to a valid integer — the port the
-       parent gateway ACTUALLY bound, exported once its TCP site is listening
-       (``dashboard.server._export_bound_port``). Below the operator override
-       (see above); above config because a bound fact from the live parent
-       beats a guess re-derived from a portless ``dashboard.url``.
-    4. Port explicitly named by ``dashboard.url`` in the config file
-       (``<data-home>/config.json``), when it parses.
-    5. The sole gateway-owned run-marker (``<data-home>/run/gateway-<port>.bin``)
-       — see :func:`_marker_port`. Skipped when no marker's port is held by a
-       verified gateway process, and refused (with a stderr hint) when several
-       are.
-    6. ``_DEFAULT_PORT`` (5476) as the final fallback.
-
-    Steps 1, 2 and 4 match the server-side ``parse_dashboard_url()`` logic so
-    that ``kirocrew token`` / ``status`` / ``logout`` / ``stop`` all hit the
-    same port the gateway is actually bound to when the user has configured a
-    non-default ``dashboard.url`` (for example a dev instance on 6777 or an
-    alternative prod port like 7778). Steps 3 and 5 cover the case where
-    nothing was configured at all but a gateway is up on a non-default port
-    (e.g. started with ``kirocrew gateway --port 6776``): the parent's own
-    export, or the running gateway's marker, is better evidence than the 5476
-    default.
-    """
-    port, _ = resolve_client_port_ex(cli_port)
-    return port
-
-
-def resolve_client_port_ex(cli_port: int | None) -> tuple[int, bool]:
-    """Like :func:`resolve_client_port`, also reporting HOW the port resolved.
-
-    The second element is ``True`` when the port came from positive evidence
-    (an explicit flag, an env var, a configured port, or a verified
-    run-marker) and ``False`` when resolution fell through to
-    ``_DEFAULT_PORT``. Callers that cache a resolution for the process
-    lifetime need the distinction: a fall-through only proves nothing was
-    discoverable *at that instant* — during gateway boot the run-marker may
-    not be written yet — so pinning it would freeze the weakest outcome
-    forever, while every positive source is stable for the process lifetime.
-    """
-    if cli_port is not None:
-        return cli_port, True
-    env_port = os.environ.get("KIROCREW_PORT")
-    if env_port:
-        try:
-            return int(env_port), True
-        except ValueError:
-            # Fall through to bound/config/marker/default — main() validates
-            # this early, but guard here too in case the helper is reached via
-            # another path.
-            pass
-    bound_port = os.environ.get("KIROCREW_BOUND_PORT")
-    if bound_port:
-        try:
-            return int(bound_port), True
-        except ValueError:
-            pass
-    cfg_port = _config_url_port()
-    if cfg_port:
-        return cfg_port, True
-    discovered = _marker_port()
-    if discovered:
-        return discovered, True
-    return _DEFAULT_PORT, False
 
 
 def _probe_dashboard_health(port: int) -> None:
@@ -661,146 +459,6 @@ def _stop(cli_port: int | None = None) -> None:
         )
         print(f"No Kiro Crew gateway currently running on port {port} (process already exited).")
         sys.exit(1)
-
-
-# Subcommands that launch a long-running KiroCrew *server* process which
-# ``kirocrew stop`` may need to terminate. These mirror the entry-point
-# subcommands dispatched in ``cli.py`` (``gateway`` / ``dashboard``; ``start``
-# is the historical alias). The task runner (``run``) is intentionally excluded:
-# it is not bound to the dashboard port, so we must never SIGTERM it from
-# ``kirocrew stop``.
-_KIROCREW_SERVER_SUBCOMMANDS = frozenset({"gateway", "dashboard", "start"})
-
-
-def _basename_stem(tok: str) -> str:
-    """Basename of *tok* without a Windows ``.exe`` suffix.
-
-    Lets the venv launchers ``python.exe`` / ``kirocrew.exe`` match the same
-    checks as their POSIX ``python`` / ``kirocrew`` counterparts. ``shlex.split``
-    with ``posix=False`` leaves quotes on some tokens, so strip them too.
-
-    Split on BOTH separators explicitly rather than via ``os.path.basename``:
-    that is host-dependent (``posixpath`` on Linux does NOT split backslashes),
-    so a Windows cmdline classified on the Linux CI fleet would keep its full
-    ``D:\\...\\kirocrew.exe`` path and never match. This is host-independent — a
-    basename is whatever follows the last ``/`` or ``\\``.
-
-    Module scope rather than nested in :func:`_args_look_like_kirocrew` so
-    :func:`_own_console_script` shares the one definition.
-    """
-    cleaned = tok.strip('"')
-    base = cleaned.replace("\\", "/").rsplit("/", 1)[-1]
-    if base.lower().endswith(".exe"):
-        base = base[:-4]
-    return base
-
-
-def _args_look_like_kirocrew(args: str) -> bool:
-    """Return ``True`` if a process command-line *args* string is a KiroCrew server.
-
-    This gates ``os.kill(pid, SIGTERM)`` in :func:`_stop`, so it must be
-    **precise** (never match an unrelated process that merely mentions
-    "kirocrew") while still recognising *every* way the gateway can be spawned.
-
-    Instead of enumerating brittle substring variants (``kiro_crew.gateway`` vs
-    ``kiro_crew gateway`` vs ``kirocrew gateway`` …), we parse the command line
-    *structurally* and key on the real module/binary name plus a known server
-    subcommand (:data:`_KIROCREW_SERVER_SUBCOMMANDS`). This is deterministic and
-    robust to interpreter path, Python version suffix, and whitespace. Two spawn
-    shapes are recognised:
-
-    * **Module invocation** — ``<python> -m kiro_crew <subcmd>`` (the form used by
-      a service install and the launchd/systemd service), plus the legacy dotted
-      form ``<python> -m kiro_crew.<subcmd>``. A Python interpreter must precede
-      ``-m`` so we don't misread some other tool's ``-m`` flag (e.g. ``grep -m``).
-    * **Console script** — ``/path/to/kirocrew <subcmd>`` (used when the
-      ``kirocrew`` wrapper resolves on ``PATH``).
-
-    Examples::
-
-        >>> _args_look_like_kirocrew("/x/python3.10 -m kiro_crew gateway")
-        True
-        >>> _args_look_like_kirocrew("python3 -m kiro_crew.dashboard")
-        True
-        >>> _args_look_like_kirocrew("/usr/local/bin/kirocrew start")
-        True
-        >>> _args_look_like_kirocrew("python -m kiro_crew run /tmp/spec.md")  # task runner
-        False
-        >>> _args_look_like_kirocrew("vim /tmp/kirocrew-notes.txt")
-        False
-    """
-    # ``ps -o args=`` (POSIX) / Win32_Process.CommandLine (Windows WMI) return a
-    # shell-style string; tokenize it the way the host shell would. On Windows
-    # use posix=False so backslash path separators survive (default posix=True
-    # eats them: ``C:\Py\python.exe`` -> ``C:Pypython.exe``, breaking the
-    # interpreter/basename checks below). Fall back to a naive split on a
-    # malformed string (e.g. an odd quote) so this best-effort check never raises.
-    try:
-        tokens = shlex.split(args, posix=not platform_compat.IS_WINDOWS)
-    except ValueError:
-        tokens = args.split()
-
-    for index, token in enumerate(tokens):
-        # --- Module form: "<python> -m kiro_crew <subcmd>" / "-m kiro_crew.<subcmd>"
-        if token == "-m" and index + 1 < len(tokens):
-            # Only treat "-m" as Python's module flag when a Python interpreter
-            # precedes it; otherwise an unrelated tool's "-m" option could be
-            # misread (e.g. "grep -m kiro_crew gateway file"). The match is
-            # case-insensitive: the macOS framework build's interpreter basename
-            # is "Python" (capital P), which a case-sensitive startswith would
-            # miss, leaving stop/restart unable to find the gateway.
-            interpreter_seen = any(
-                _basename_stem(t).lower().startswith("python") for t in tokens[:index]
-            )
-            if interpreter_seen:
-                # "kiro_crew.gateway" -> ("kiro_crew", "gateway"); a bare
-                # "kiro_crew" -> ("kiro_crew", "").
-                package, _, dotted_subcmd = tokens[index + 1].partition(".")
-                if package == "kiro_crew":
-                    # Dotted submodule form: ``-m kiro_crew.gateway``.
-                    if dotted_subcmd in _KIROCREW_SERVER_SUBCOMMANDS:
-                        return True
-                    # Subcommand-as-argument form: ``-m kiro_crew gateway``. The
-                    # subcommand is argparse's first positional after the module,
-                    # i.e. always at index+2. Check only that slot so a later
-                    # positional/flag value cannot match — e.g.
-                    # ``-m kiro_crew run gateway`` ("gateway" is a file argument
-                    # to the task runner) must NOT be treated as a server.
-                    if (
-                        index + 2 < len(tokens)
-                        and tokens[index + 2] in _KIROCREW_SERVER_SUBCOMMANDS
-                    ):
-                        return True
-
-        # --- Console-script form: ".../kirocrew <subcmd>" (or kirocrew.exe on Win)
-        if (
-            _basename_stem(token) == "kirocrew"
-            and index + 1 < len(tokens)
-            and tokens[index + 1] in _KIROCREW_SERVER_SUBCOMMANDS
-        ):
-            return True
-
-    return False
-
-
-def _is_kirocrew_process(pid: int) -> bool:
-    """Return ``True`` if *pid* looks like a KiroCrew gateway process.
-
-    Resolves the process command line cross-platform via
-    :func:`platform_compat.process_command_line` (Linux ``/proc``, macOS ``ps``,
-    Windows ``Win32_Process`` WMI — the venv ``kirocrew.exe`` re-execs
-    ``python.exe`` so the image name alone is ambiguous there) and defers
-    classification to :func:`_args_look_like_kirocrew`.
-
-    ``process_command_line`` returns ``""`` on any failure (dead PID, missing
-    ``ps``, WMI error), which classifies as "not a match" — _stop()'s separate
-    ``listening_pid_tool_available()`` check already surfaces the tool-absent
-    case, so this never needs to raise.
-    """
-    out = platform_compat.process_command_line(pid)
-    if not out:
-        return False
-    return _args_look_like_kirocrew(out)
 
 
 def _pid_exited(pid: int) -> bool:
